refactor: split navigation sidebar
